Merge into local branch

+++ src/core/StringBuilder.hpp
@@ -47,6 +47,8 @@ public:
 	/**
 	 * Appends the given string of the given length to the current string. Essentially used to ignore null terminators or copy
 	 * the data faster as the length is already known.
+	 * @param text Pointer to the UTF-8 text to append.
+	 * @param textLength Number of bytes to copy. (Can be used to append single bytes rather than codepoints)
 	 */
 	void Append(const utf8 *text, int textLength) {
 		EnsureCapacity(_length + textLength + 1);
@@ -123,7 +125,7 @@ private:
 	{
 		if (_capacity > capacity) return;
 
-		_capacity = Math::Max(8U, _capacity);
+		_capacity = Math::Max((size_t)8, _capacity);
 		while (_capacity < capacity) {
 			_capacity *= 2;
 		}
